requirements.txt
dill
mdutils
pycm
pytorch
pytorch-nlp
pytorch-lightning
sentencepiece
